edited according to comments in PR
