+++ core/src/test/scala/examples/DummyConsumer.scala
@@ -1,32 +1,18 @@
 package examples
 
 import akka.actor.ActorSystem
-import akka.stream.scaladsl.{Flow, GraphDSL, Sink, Source}
+import akka.stream.scaladsl.{Flow, GraphDSL, Source}
 import akka.stream.{ActorMaterializer, ActorMaterializerSettings, SourceShape}
 import com.typesafe.scalalogging.slf4j.LazyLogging
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.apache.kafka.common.serialization.{ByteArrayDeserializer, StringDeserializer}
 
 import scala.concurrent.Await
-import scala.util.Failure
 import com.softwaremill.react.kafka2._
 import scala.concurrent.duration._
 
-object Streams {
-  def shutdownAsOnComplete[T](implicit as: ActorSystem) = Sink.onComplete[T] {
-    case Failure(ex) =>
-      println("Stream finished with error")
-      ex.printStackTrace()
-      as.terminate()
-      println("Terminate AS.")
-    case _ =>
-      println("Stream finished successfully")
-      as.terminate()
-      println("Terminate AS.")
-  }
-}
-
-// tbd. What is the purpose of this?
+// Reads from topic "dummy", processes messages and commits offset into kafka after processing.
+// This provides at-least-once delivery guarantee. Also, shows how to perform graceful shutdown.
 //
 // Usage:
 //    sbt core/test:run
@@ -39,7 +25,11 @@ object DummyConsumer extends App with LazyLogging {
       .withInputBuffer(16, 16)
   )
 
-  val provider = ConsumerProvider("localhost:9092", new ByteArrayDeserializer, new StringDeserializer)
+  val provider = ConsumerProvider(
+    "localhost:9092",
+    new ByteArrayDeserializer,
+    new StringDeserializer
+  )
     .setup(TopicSubscription("dummy"))
     .groupId("c5")
     .autoCommit(false)
@@ -57,7 +47,7 @@ object DummyConsumer extends App with LazyLogging {
   val control =
     Source.fromGraph(graph)
       .mapAsync(8)(identity)
-      .to(Streams.shutdownAsOnComplete)
+      .to(shutdownAsOnComplete)
       .run()
 
   sys.addShutdownHook {
